[#214] Move "reader settings" util functions into new file
