+++ src/components/navbar/ReaderNavBar.tsx
@@ -28,7 +28,6 @@ import Collapse from '@mui/material/Collapse';
 import Button from '@mui/material/Button';
 import { styled } from '@mui/system';
 import useBackTo from 'util/useBackTo';
-import { getMetadataFrom } from 'util/metadata';
 
 const Root = styled('div')(({ theme }) => ({
     top: 0,
@@ -105,27 +104,6 @@ const OpenDrawerButton = styled(IconButton)(({ theme }) => ({
         backgroundColor: theme.palette.custom.light,
     },
 }));
-
-const defaultReaderSettings = () => ({
-    staticNav: false,
-    showPageNumber: true,
-    continuesPageGap: false,
-    loadNextonEnding: false,
-    readerType: 'ContinuesVertical',
-} as IReaderSettings);
-
-const getReaderSettingsFromMetadata = (
-    meta?: IMetadata,
-): IReaderSettings => ({
-    ...getMetadataFrom(
-        { meta },
-        Object.entries(defaultReaderSettings()) as MetadataKeyValuePair[],
-    ) as unknown as IReaderSettings,
-});
-
-export const getReaderSettingsFor = (
-    { meta }: IMangaCard | IManga,
-): IReaderSettings => getReaderSettingsFromMetadata(meta);
 
 interface IProps {
     settings: IReaderSettings
